refactor: 코드래빗 피드백 반영
- consumer NPE 방지 로직 추가
- package 명칭 오타 수정
- 이벤트 발행 타임아웃 설정
